refactor(ResponseUtil): 헤더에 위치한 Location을 뽑아내는 함수 생성

+++ app/src/main/java/woowacourse/shopping/data/datasource/basket/BasketRemoteDataSource.kt
@@ -7,7 +7,7 @@ interface BasketRemoteDataSource {
 
     fun getAll(): Result<List<BasketProductEntity>>
 
-    fun add(product: ProductEntity): Result<Int>
+    fun add(product: ProductEntity): Result<Long>
 
     fun update(basketProduct: BasketProductEntity): Result<Unit>
